Game: prepare to add particles
Split out the implicit packed array implementation into a
couple of templated functions so they can be used on more than
just projectiles.

Add some structures to prepare for particle system implemetation.
The particles will be mostly stateless except for things like position/rotation.
A ParticleEmitter emits particles, which have an index for a "template"
which is the properties of that particular particle (with keyframes for values animated over lifetime).
ParticleEffect structs hold multiple emitters.